+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- test(): stroke projection [#9110](https://github.com/fabricjs/fabric.js/pull/9110)
 - fix(SVGParser): avoid crashing on SVG that use @import css feature [#9602](https://github.com/fabricjs/fabric.js/pull/9602)
 - fix(): compositionEnd event handler is not registered correctly. (regression from f91362c ) [#9610](https://github.com/fabricjs/fabric.js/pull/9610)
 - ci(): Add a test case from the multiple selection use case for groups [#9599](https://github.com/fabricjs/fabric.js/pull/9599)